+++ protocols/kad/src/behaviour.rs
@@ -23,8 +23,8 @@
 mod test;
 
 use crate::addresses::Addresses;
+use crate::bootstrap;
 use crate::handler::{Handler, HandlerEvent, HandlerIn, RequestId};
-use crate::jobs::*;
 use crate::kbucket::{self, Distance, KBucketsTable, NodeStatus};
 use crate::protocol::{ConnectionType, KadPeer, ProtocolConfig};
 use crate::query::{Query, QueryConfig, QueryId, QueryPool, QueryPoolState};
@@ -33,7 +33,8 @@ use crate::record::{
     store::{self, RecordStore},
     ProviderRecord, Record,
 };
-use crate::{bootstrap, K_VALUE};
+use crate::K_VALUE;
+use crate::{jobs::*, protocol};
 use fnv::{FnvHashMap, FnvHashSet};
 use instant::Instant;
 use libp2p_core::{ConnectedPoint, Endpoint, Multiaddr};
@@ -189,22 +190,11 @@ pub struct Config {
 }
 
 impl Default for Config {
+    /// Returns the default configuration.
+    ///
+    /// Deprecated: use `Config::new` instead.
     fn default() -> Self {
-        Config {
-            kbucket_pending_timeout: Duration::from_secs(60),
-            query_config: QueryConfig::default(),
-            protocol_config: Default::default(),
-            record_ttl: Some(Duration::from_secs(36 * 60 * 60)),
-            record_replication_interval: Some(Duration::from_secs(60 * 60)),
-            record_publication_interval: Some(Duration::from_secs(24 * 60 * 60)),
-            record_filtering: StoreInserts::Unfiltered,
-            provider_publication_interval: Some(Duration::from_secs(12 * 60 * 60)),
-            provider_record_ttl: Some(Duration::from_secs(24 * 60 * 60)),
-            kbucket_inserts: BucketInserts::OnConnected,
-            caching: Caching::Enabled { max_peers: 1 },
-            periodic_bootstrap_interval: Some(Duration::from_secs(5 * 60)),
-            automatic_bootstrap_throttle: Some(bootstrap::DEFAULT_AUTOMATIC_THROTTLE),
-        }
+        Self::new(protocol::DEFAULT_PROTO_NAME)
     }
 }
 
@@ -224,6 +214,32 @@ pub enum Caching {
 }
 
 impl Config {
+    /// Builds a new `Config` with the given protocol name.
+    pub fn new(protocol_name: StreamProtocol) -> Self {
+        Config {
+            kbucket_pending_timeout: Duration::from_secs(60),
+            query_config: QueryConfig::default(),
+            protocol_config: ProtocolConfig::new(protocol_name),
+            record_ttl: Some(Duration::from_secs(36 * 60 * 60)),
+            record_replication_interval: Some(Duration::from_secs(60 * 60)),
+            record_publication_interval: Some(Duration::from_secs(24 * 60 * 60)),
+            record_filtering: StoreInserts::Unfiltered,
+            provider_publication_interval: Some(Duration::from_secs(12 * 60 * 60)),
+            provider_record_ttl: Some(Duration::from_secs(24 * 60 * 60)),
+            kbucket_inserts: BucketInserts::OnConnected,
+            caching: Caching::Enabled { max_peers: 1 },
+            periodic_bootstrap_interval: Some(Duration::from_secs(5 * 60)),
+            automatic_bootstrap_throttle: Some(bootstrap::DEFAULT_AUTOMATIC_THROTTLE),
+        }
+    }
+
+    /// Returns the default configuration.
+    #[deprecated(note = "Use `Config::new` instead")]
+    #[allow(clippy::should_implement_trait)]
+    pub fn default() -> Self {
+        Default::default()
+    }
+
     /// Sets custom protocol names.
     ///
     /// Kademlia nodes only communicate with other nodes using the same protocol
@@ -233,6 +249,8 @@ impl Config {
     /// More than one protocol name can be supplied. In this case the node will
     /// be able to talk to other nodes supporting any of the provided names.
     /// Multiple names must be used with caution to avoid network partitioning.
+    #[deprecated(note = "Use `Config::new` instead")]
+    #[allow(deprecated)]
     pub fn set_protocol_names(&mut self, names: Vec<StreamProtocol>) -> &mut Self {
         self.protocol_config.set_protocol_names(names);
         self